vmae-vitb

+++ alphaction/dataset/datasets/evaluation/ava/ava_eval.py
@@ -32,7 +32,7 @@ def do_ava_evaluation(dataset, predictions, output_folder, logger):
 
 def make_image_key(video_id, timestamp):
     """Returns a unique identifier for a video id & timestamp."""
-    return "%s,%04d" % (video_id, int(timestamp))
+    return "%s,%04d" % (video_id, float(timestamp))
 
 
 def decode_image_key(image_key):
@@ -235,3 +235,4 @@ def evaluate_predictions_on_ava(eval_file_paths, ava_results, csv_result_file, l
     metrics = pascal_evaluator.evaluate()
     print_time(logger, "run_evaluator", start)
     return metrics
+
